cccnext-apply-ml-services-5b8d067f245d/ccctc_spam/requirements.txt
pandas==0.21.0
tabulate==0.8.2
scikit-learn==0.19.1
scipy==1.0.0
pyyaml==3.12
category_encoders==1.2.4
sklearn-pandas==1.6.0
xgboost==0.6a2
matplotlib==2.0.2
flask==0.12
python-dotenv==0.6.1
werkzeug==0.11.15
jwcrypto==0.4.2
gunicorn==19.8.1
